# swagger.yml
openapi: 3.0.0
info:
  title: Mailing Manager
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Local development server
paths:
...
  /mail/{role}:
    post:
      summary: Get mails for role
      description: Return list of mails for role using email
      tags:
        - Mail
      parameters:
        - name: role
          in: path
          required: true
          description: The role for which to get mails
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
              required:
                - email
      responses:
        '200':
          description: Success
...
  /mail/add:
    post:
      summary: Add new mail
      description: Add new mail to db
      tags:
        - Mail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                theme:
                  type: string
                target:
                  type: string
                community:
                  type: string
                template:
                  type: string
                status:
                  type: string
                analyst:
                  type: integer
                product:
                  type: integer
                editor:
                  type: integer
                main_editor:
                  type: integer
                marketolog:
                  type: integer
              required:
                - theme
                - target
      responses:
        '200':
          description: Success